Introduce the `ortiofay` CLI
This CLI implements a load testing harness similar to strest-grpc,
slow_cooker, and fortio. The CLI provies `load` and `server`
subcommands. The load generator implements rate limiting and concurrent
load generation. It reports results vi an admin endpoint (in a
fortio-like report structure).

There is an additional `controller` subcommand that is disabled by
default. It is intended to, eventually, drive test setups in Kubernetes
(driven by CRD). It doesn't do all that much yet, so it is currently
disabled by feature-flag.

This is a work in progress. It is basically functional, though deficient
(and possibly buggy), especially with regartd to reporting results.